Transform abbreviateNumber into a number utils function
