+++ packages/shared-components/PostStats/index.jsx
@@ -13,6 +13,9 @@ import {
 import { 
   constants,
 } from '@bufferapp/publish-utils';
+import { 
+  abbreviateNumber,
+} from '@bufferapp/publish-utils/number';
 
 const statsBarStyle = {
   display: 'flex',
@@ -29,26 +32,6 @@ const statsCellStyle = {
   borderTop: `${borderWidth} solid ${mystic}`,
   padding: '0.5rem',
   borderRight: `${borderWidth} solid ${mystic}`,
-};
-
-/* Abbreviates numbers, eg: if we want 2 decimal places, it displays 1240 as 1.24k */
-const abbreviateNumber = (number, decPlaces) => {
-  decPlaces = 10 ** decPlaces;
-  const abbreviations = [ "k", "m", "b", "t" ];
-
-  for (let i=abbreviations.length-1; i>=0; i--) {
-    const size = 10 ** ((i + 1) * 3);
-    if(size <= number) {
-      number = Math.round(number*decPlaces/size)/decPlaces;
-      if((number == 1000) && (i < abbreviations.length - 1)) {
-        number = 1;
-        i++;
-      }
-      number += abbreviations[i];
-      break;
-    }
-  }
-  return number;
 };
 
 const PostStats = ({
